Refactor: date formatting 추상화
